pyproject.toml
[tool.poetry]
name = "slackbot"
version = "0.1.0"
description = ""
authors = ["engkimo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
openai = "^1.7.2"
async-timeout = "^4.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
